Bufgix regarding the stable Paper versions
Eue to an error, only unstable versions of paper have been considered so far. Now it is possible to include or exclude the stable versions of Paper in the config.yml.

A new line has been added to the config for this, which must either be inserted manually or the plugin folder must first be deleted again and regenerated independently.